Remove vars.Report from all but metrics plugin
This should only happen there.

Fixes: #2717

It depends (and conflicts) with #2715
