+++ plugin/metrics/context.go
@@ -3,7 +3,7 @@ package metrics
 import (
 	"context"
 
-	"github.com/coredns/coredns/plugin/metrics/vars"
+	"github.com/coredns/coredns/plugin"
 )
 
 // WithServer returns the current server handling the request. It returns the
@@ -15,4 +15,10 @@ import (
 // Basic usage with a metric:
 //
 // <metric>.WithLabelValues(metrics.WithServer(ctx), labels..).Add(1)
-func WithServer(ctx context.Context) string { return vars.WithServer(ctx) }
+func WithServer(ctx context.Context) string {
+	srv := ctx.Value(plugin.ServerCtx{})
+	if srv == nil {
+		return ""
+	}
+	return srv.(string)
+}
